Some refactoring and issue #105
