+++ src/Message/ServerNotifyRequest.php
@@ -19,6 +19,18 @@ class ServerNotifyRequest extends AbstractRequest implements NotificationInterfa
     use ServerNotifyTrait;
 
     /**
+     * Valid status responses.
+     */
+    const RESPONSE_STATUS_OK        = 'OK';
+    const RESPONSE_STATUS_ERROR     = 'ERROR';
+    const RESPONSE_STATUS_INVALID   = 'INVALID';
+
+    /**
+     * Live separator for return message to Sage Pay.
+     */
+    const LINE_SEP = "\r\n";
+
+    /**
      * Copy of the POST data sent in.
      */
     protected $data;
@@ -43,7 +55,7 @@ class ServerNotifyRequest extends AbstractRequest implements NotificationInterfa
     }
 
     /**
-     * Set the saved TransactionReference.
+     * Legacy support.
      * We are only interested in extracting the security key here.
      * It makes more sense to use setSecurityKey().
      *
@@ -51,9 +63,9 @@ class ServerNotifyRequest extends AbstractRequest implements NotificationInterfa
      */
     public function setTransactionReference($reference)
     {
-        // Is this a JSON string?
         if (strpos($reference, 'SecurityKey') !== false) {
-            // Yes. Decode it then extact the security key.
+            // A JSON string provided - the legacy transactionReference format.
+            // Decode it then extact the securityKey.
             // We only need the security key here for the signature; all other
             // items from the reference will be in the server request.
 
@@ -63,24 +75,24 @@ class ServerNotifyRequest extends AbstractRequest implements NotificationInterfa
                 $this->setSecurityKey($parts['SecurityKey']);
             }
         }
+
+        return $this;
     }
 
     /**
-     * Get the Sage Pay Responder.
+     * Legacy support.
      *
-     * @param string $data message body.
-     * @return ServerNotifyResponse
+     * @param mixed $data ignored
+     * @return $this
      */
     public function sendData($data)
     {
-        $data['vendor'] = $this->getVendor();
-        $data['securityKey'] = $this->getSecurityKey();
-
-        return $this->response = new ServerNotifyResponse($this, $data);
+        return $this;
     }
 
     /**
      * Set the SecurityKey that we saved locally.
+     * This is our one-time secret for the signature hash.
      *
      * @return self
      */
@@ -92,5 +104,114 @@ class ServerNotifyRequest extends AbstractRequest implements NotificationInterfa
     public function getSecurityKey()
     {
         return $this->getParameter('SecurityKey');
+    }
+
+    /**
+     * Confirm
+     *
+     * Notify Sage Pay you received the payment details and wish to confirm the payment.
+     *
+     * @param string URL to forward the customer to.
+     * @param string Optional human readable reasons for accepting the transaction.
+     */
+    public function confirm($nextUrl, $detail = null)
+    {
+        // If the signature is invalid, then do not allow the confirm.
+        if (! $this->isValid()) {
+            throw new InvalidResponseException('Cannot confirm an invalid notification');
+        }
+
+        $this->sendResponse(static::RESPONSE_STATUS_OK, $nextUrl, $detail);
+    }
+
+    /**
+     * Alias for confirm(), trying to define some more general conventions.
+     */
+    public function accept($nextUrl, $detail = null)
+    {
+        return $this->confirm($nextUrl, $detail);
+    }
+
+    /**
+     * Error
+     *
+     * Notify Sage Pay you received the payment details but there was an error and the payment
+     * cannot be completed.
+     *
+     * @param string URL to foward the customer to.
+     * @param string Optional human readable reasons for not accepting the transaction.
+     */
+    public function error($nextUrl, $detail = null)
+    {
+        // If the signature is invalid, then do not allow the reject.
+        // CHECKME: why?
+        if (! $this->isValid()) {
+            throw new InvalidResponseException('Cannot reject an invalid notification');
+        }
+
+        $this->sendResponse(static::RESPONSE_STATUS_ERROR, $nextUrl, $detail);
+    }
+
+    /**
+     * Alias for error(), trying to define some more general conventions.
+     */
+    public function reject($nextUrl, $detail = null)
+    {
+        return $this->error($nextUrl, $detail);
+    }
+
+    /**
+     * Invalid
+     *
+     * Notify Sage Pay you received *something* but the details were invalid and no payment
+     * cannot be completed. Invalid should be called if you are not happy with the contents
+     * of the POST, such as the MD5 hash signatures did not match or you do not wish to proceed
+     * with the order.
+     *
+     * @param string URL to foward the customer to.
+     * @param string Optional human readable reasons for not accepting the transaction.
+     */
+    public function invalid($nextUrl, $detail = null)
+    {
+        $this->sendResponse(static::RESPONSE_STATUS_INVALID, $nextUrl, $detail);
+    }
+
+    /**
+     * Construct the response body.
+     *
+     * @param string The status to send to Sage Pay, one of static::RESPONSE_STATUS_*
+     * @param string URL to forward the customer to.
+     * @param string Optional human readable reason for this response.
+     */
+    public function getResponseBody($status, $nextUrl, $detail = null)
+    {
+        $body = array(
+            'Status=' . $status,
+            'RedirectUrl=' . $nextUrl,
+        );
+
+        if ($detail !== null) {
+            $body[] = 'StatusDetail=' . $detail;
+        }
+
+        return implode(static::LINE_SEP, $body);
+    }
+
+    /**
+     * Respond to SagePay confirming or rejecting the notification.
+     *
+     * @param string The status to send to Sage Pay, one of static::RESPONSE_STATUS_*
+     * @param string URL to forward the customer to.
+     * @param string Optional human readable reason for this response.
+     */
+    public function sendResponse($status, $nextUrl, $detail = null)
+    {
+        $message = $this->getResponseBody($status, $nextUrl, $detail);
+
+        echo $message;
+
+        if ((bool)$this->getExitOnResponse()) {
+            exit;
+        }
     }
 }
